requirements.txt
asn1crypto==0.24.0
backports-abc==0.5
beautifulsoup4==4.6.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
enum34==1.1.6
future==0.17.1
futures==3.3.0
html5lib==0.999999999
idna==2.6
ipaddress==1.0.22
keyring==10.6.0
keyrings.alt==3.0
lxml==4.2.1
olefile==0.45.1
PAM==0.4.2
pexpect==4.2.1
Pillow==5.1.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
pyinotify==0.9.6
python-telegram-bot==12.0.0b1
pyxdg==0.25
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.12.0
tornado==5.1.1
webencodings==0.5
